==> 00_data/00_scripts/1_VHDB_screen.ipynb <==
import pandas as pd
df = pd.read_table('../virushostdb.daily.tsv')
df.columns.values

筛选phage-host条目

column_list= ['virus name', 'refseq id', 'virus tax id', 'virus lineage', 'evidence',
                'host name', 'host tax id', 'host lineage']

df2 = df[column_list].copy()

df2.columns = ['virus_name', 'refseq_id', 'virus_taxid', 'virus_lineage', 'evidence',
                'host_name', 'host_taxid', 'host_lineage']


df3 = df2.dropna(axis=0, how='any')


def select_phage(x):
    """
    Selects phage  from all viruses.
    """
    try:
        kingdom = x.split(';')[0].strip()
        if kingdom == 'Bacteria' or kingdom == 'Archaea':
            return 1
        elif kingdom == 'Eukaryota':
            return 0
    except:
        print(x)


df_phage = df3[df3['host_lineage'].apply(select_phage) == 1].copy()

df_phage['host_taxid'] = df_phage['host_taxid'].astype(int)
df_phage.to_csv('../virushostdb.daily.phage.tsv', sep='\t', index=False)


发现24个phage有多个id


df_phage_multiple=df_phage[df_phage['refseq_id'].str.contains(',')]
df_phage_multiple.to_csv('../virushostdb.daily.phage.multiple.tsv', sep='\t', index=False)

len(set(df_phage_multiple['virus_name']))

检查后发现只有9个phage不完整，其他的只是同一个名称的不同测序结果，对不完整的phage直接进行删除


# 过滤不完整病毒,9个
uncompleted_virus = ['Bacillus phage SPG24', 'Pseudomonas phage phi12', 'Pseudomonas phage phi13', 
'Pseudomonas phage phi2954','Pseudomonas phage phi8', 'Pseudomonas phage phiNN', 
'Pseudomonas phage phiYY', 'Pseudomonas virus phi6', 'Salmonella phage SP069']

df_completed = df_phage[~df_phage['virus_name'].isin(uncompleted_virus)]
print(df_phage.shape)
print(df_completed.shape)
df_phage=df_completed.copy()

# 保存
df_phage.to_csv('../virushostdb.daily.phage.tsv', sep='\t', index=False)


VHDB phage name有重复的是因为一个phage可能记录了多个host

df_phage['virus_name'].value_counts()


## 添加 tax lineage信息
根据host taxid生成tax lineage信息，只保留界门纲目科属种，如果不在有缺失则用上一级+unclassified

# cd ../
# mkdir -p temp/ lineage/
# !cut -f 7 virushostdb.daily.phage.tsv > temp/taxid.tsv
# !cat temp/taxid.tsv | taxonkit lineage | taxonkit reformat -f "{k}\t{p}\t{c}\t{o}\t{f}\t{g}\t{s}\t{t}" -P  | cut -f 3- |  tr '\t' ';' | tee temp/lineage.tsv

# !cat temp/taxid.tsv | taxonkit lineage -r -L | tee temp/rank.tsv
# !paste temp/rank.tsv temp/lineage.tsv | tail -n+2 > lineage/lineage_host.tsv
# !rm temp/taxid.tsv temp/lineage.tsv temp/rank.tsv


根据virus taxid生成界门纲目科属种的物种注释

# !cut -f 3 virushostdb.daily.phage.tsv > temp/taxid.tsv
# !cat temp/taxid.tsv | taxonkit lineage | taxonkit reformat -f "{k}\t{p}\t{c}\t{o}\t{f}\t{g}\t{s}\t{t}" -P  | cut -f 3- |  tr '\t' ';' | tee lineage/lineage_virus.tsv




lineage_host = pd.read_table('../lineage/lineage_host.tsv', header=None,
                        names=['host_taxid', 'host_rank', 'host_lineage'])
lineage_host

lineage_virus = pd.read_table('../lineage/lineage_virus.tsv', header=None,
                             names=['virus_lineage'])
lineage_virus


df_phage['host_rank'] = lineage_host['host_rank'].to_list()

df_phage['host_lineage'] =lineage_host['host_lineage'].to_list()
df_phage['virus_lineage'] = lineage_virus['virus_lineage'].to_list()
df_phage=df_phage.reset_index(drop=True)
df_phage.to_csv('../virushostdb.daily.phage.tsv', sep='\t', index=False)


df_phage


df_phage['host_rank'].value_counts()

## 过滤宿主名有问题的两个
#{'bacterium', 'unidentified bacterial endosymbiont'}
unidentified_host = ['bacterium', 'unidentified bacterial endosymbiont']

df_phage = df_phage[~df_phage['host_name'].isin(unidentified_host)]
df_phage.to_csv('../virushostdb.daily.phage.tsv',
                   sep='\t', index=False)


# 过滤不完整病毒,9个
under_species = ['species', 'strain', 'no rank',
                'serotype', 'subspecies', 'serogroup',
                'forma specialis', 'isolate']

df_filtered = df_phage[df_phage['host_rank'].isin(under_species)]
print(df_phage.shape)
print(df_filtered.shape)
df_filtered.to_csv('../virushostdb.daily.phage(filtered).tsv',
                sep='\t', index=False)


##  统计信息

print(len(df_phage))
print(len(set(df_phage['host_name'])))
print(len(set(df_phage['virus_name'])))

print(len(df_filtered))
print(len(set(df_filtered['host_name'])))
print(len(set(df_filtered['virus_name'])))
