==> 00_data/00_scripts/5_filter_gcf.ipynb <==
import pandas as pd

## 给output.tsv添加find物种注释


df2 = pd.read_table(
    '../virushostdb_output.tsv', sep='\t')

df_findtax = pd.read_table('../lineage/find_info.tsv', sep='\t',header=None, dtype=object)
# del df2['find_lineage']
df_findtax.columns = [ 'find_name','find_taxid', 'find_taxrank','find_lineage' ]
df2['find_taxid'] = df_findtax['find_taxid']
df2['find_taxrank'] = df_findtax['find_taxrank']
df2['find_lineage'] = df_findtax['find_lineage']
df2.to_csv('../virushostdb_output.tsv', sep='\t', index=False)

## 过滤

df2 = pd.read_table('../virushostdb_output.tsv', sep='\t')
df3 = df2[df2['find_name'] == df2['report_host']]
len(df3)

len(set(df2['report_host']))

df2['find_taxrank'].value_counts()

df2 = df2[~(df2['find_taxrank'] == 'genus')]
df2 = df2[~(df2['find_taxrank'] == 'family')]
df2 = df2[~(df2['find_taxrank'] == 'species group')]
df2 = df2[~(df2['find_taxrank'] == 'class')]
df2 = df2[~(df2['find_taxrank'] == 'order')]


df2['find_taxrank'].value_counts()

print(len(df2))
print(len(set(df2['report_host'])))
print(len(set(df2['find_name'])))
print(len(set(df2['new_gcf'])))

df2.to_csv('../virushostdb_output(filtered).tsv', sep='\t', index=False)
