==> 00_data/00_scripts/6_gcf_json.ipynb <==
根据phage-host信息生成host_json文件

from tqdm import tqdm
import json
input_f = '../virushostdb_output(filtered).tsv'
host_dict = {}
num_lines = sum(1 for line in open(input_f, 'r'))-1
with open(input_f,mode='r',encoding='utf-8') as f:
    f.readline()
    for line in tqdm(f,total=num_lines):
        line_list=line.strip().split('\t')
        # print(len(line_list))
        find_lineage=line_list[16]
        find_taxrank=line_list[15]
        find_taxid=line_list[14]
        assembly_level=line_list[13]
        assembly_category=line_list[12]
        new_gcf=line_list[11]
        gcfs=line_list[10]
        find_name=line_list[9]
        host_tax_rank=line_list[8]
        host_lineage=line_list[7]
        host_taxid=line_list[6]
        report_host=line_list[5]
        if report_host not in host_dict:
            host_dict[report_host] = {
                'host_taxid': host_taxid,
                'host_tax_rank': host_tax_rank,
                'host_lineage': host_lineage,
                'find_name':find_name,
                'gcfs':gcfs,
                'new_gcf':new_gcf,
                'assembly_category':assembly_category,
                'assembly_level':assembly_level,
                'find_taxid':find_taxid,
                'find_taxrank':find_taxrank,
                'find_lineage':find_lineage,
                }
js = json.dumps(host_dict,indent=4)



import pandas as pd
table=[]
line_dict={}
for host_name in host_dict.keys():
    line_dict={"host_name":host_name}
    line_dict.update(host_dict[host_name])
    table.append(line_dict)
    
df=pd.DataFrame(table)



len(df)

## 构建gcf json文件

df_gcf=df[["new_gcf", "find_name", 'find_lineage',"assembly_category", "assembly_level"]]


gcf_dict = {}

for i in range(df_gcf.shape[0]):
    gcf = df_gcf.iloc[i, 0]
    if gcf not in gcf_dict:
        gcf_dict[gcf]={
            'find_name':[df_gcf.iloc[i,1]],
            'find_lineage':[df_gcf.iloc[i,2]],
            'assembly_category':df_gcf.iloc[i,3],
            'assembly_level':df_gcf.iloc[i,4],
        }
    else:
        find_name=df_gcf.iloc[i,1]
        if find_name not in gcf_dict[gcf]['find_name']:
            gcf_dict[gcf]['find_name'].append(find_name)

js = json.dumps(gcf_dict, indent=4)
fp = open('../find_gcf.json', 'w')
fp.write(js)
fp.close()


len(gcf_dict)
